+++ src/rules/core/strict.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   rules: {
     strict: ["error", "never"], // typescript inserts `'use strict';` for us
   },