+++ src/rules/functional/currying.js
@@ -1,5 +1,7 @@
-module.exports = {
-  plugins: ["functional"],
+import functionalPlugin from "eslint-plugin-functional";
+
+export default {
+  plugins: { functional: functionalPlugin },
   rules: {
     "functional/functional-parameters": "off", // TODO: Enable?
   },